Add support for staged install
This is a general refactor of the makefile with the introduction of
staged install.

Is now possible to install with make install DESTDIR=build
and it will build the default xlxd install structure under the build
folder.

This if extremely usefull for building deb packages without using root
user and without interfeering with the host xlxd installation.

It also define some standard make directory variables, all defaulted
to /xlxd, to better customize the software at compile time.

make staged install:
https://www.gnu.org/prep/standards/html_node/DESTDIR.html
make directory variables:
https://www.gnu.org/software/make/manual/html_node/Directory-Variables.html