isdigit() may return any non-zero result for success
POSIX states that the return value of isdigit(): "shall return non-zero
if c is a decimal digit; otherwise, they shall return 0."

Thus the form:

  ok &= isdigit(x)

is invalid since the runtime is not required to return 1.

This bug was observed on Debian 11 while using the clang toolchain.
In that environment, isdigit() returns 2048 for a positive match.